DMC-918 test rework: update test after review
